backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
litellm
daisys
pydantic
python-multipart
aiofiles